refactor(suite): stake file naming
