+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/UnstakeModal/UnstakeForm/UnstakeButton.tsx
@@ -1,46 +1,48 @@
 import { selectHasRunningDiscovery } from '@suite-common/wallet-core';
 import type { SelectedAccountLoaded } from '@suite-common/wallet-types';
-import { Modal, Tooltip } from '@trezor/components';
+import { Button, Tooltip } from '@trezor/components';
 import { EventType, analytics } from '@trezor/suite-analytics';
 
 import { Translation } from 'src/components/suite';
 import { useDevice, useSelector } from 'src/hooks/suite';
 import { useMessageSystemStaking } from 'src/hooks/suite/useMessageSystemStaking';
-import { useStakeEthFormContext } from 'src/hooks/wallet/useStakeEthForm';
+import { useUnstakeFormContext } from 'src/hooks/wallet/useUnstakeForm';
 import { CRYPTO_INPUT, FIAT_INPUT } from 'src/types/wallet/stakeForms';
 
-export const StakeButton = () => {
+export const UnstakeButton = () => {
     const { device, isLocked } = useDevice();
     const selectedAccount = useSelector(
         state => state.wallet.selectedAccount,
     ) as SelectedAccountLoaded;
-    const {
-        onSubmit,
-        handleSubmit,
-        formState: { errors, isSubmitting },
-        isComposing,
-        watch,
-        currency,
-    } = useStakeEthFormContext();
-    const { isStakingDisabled, stakingMessageContent } = useMessageSystemStaking(
+    const { isUnstakingDisabled, unstakingMessageContent } = useMessageSystemStaking(
         selectedAccount.network.symbol,
     );
-    const isDiscoveryRunning = useSelector(selectHasRunningDiscovery);
+
+    const {
+        isComposing,
+        formState: { isSubmitting, errors },
+        handleSubmit,
+        watch,
+        signTx,
+        currency,
+    } = useUnstakeFormContext();
 
     const hasValues = Boolean(watch(FIAT_INPUT) || watch(CRYPTO_INPUT));
     // used instead of formState.isValid, which is sometimes returning false even if there are no errors
     const formIsValid = Object.keys(errors).length === 0;
+
     const isDisabled =
         !(formIsValid && hasValues) || isSubmitting || isLocked() || !device?.available;
+    const isDiscoveryRunning = useSelector(selectHasRunningDiscovery);
 
-    const onStakeClick = () => {
-        handleSubmit(onSubmit)();
+    const onUnstakeClick = () => {
+        handleSubmit(signTx)();
 
         analytics.report({
-            type: EventType.StakingStake,
+            type: EventType.StakingUnstake,
             payload: {
                 action: 'continue',
-                step: 'stake-form-modal',
+                step: 'unstake-form-modal',
                 currency,
                 networkSymbol: selectedAccount.account.symbol,
             },
@@ -48,15 +50,16 @@ export const StakeButton = () => {
     };
 
     return (
-        <Tooltip content={stakingMessageContent}>
-            <Modal.Button
-                isDisabled={isDisabled || isStakingDisabled}
+        <Tooltip content={unstakingMessageContent}>
+            <Button
+                type="submit"
+                isDisabled={isDisabled || isUnstakingDisabled}
                 isLoading={isComposing || isSubmitting || isDiscoveryRunning}
-                onClick={onStakeClick}
-                icon={isStakingDisabled ? 'info' : undefined}
+                onClick={onUnstakeClick}
+                icon={isUnstakingDisabled ? 'info' : undefined}
             >
-                <Translation id="TR_CONTINUE" />
-            </Modal.Button>
+                <Translation id="TR_STAKE_UNSTAKE" />
+            </Button>
         </Tooltip>
     );
 };
