+++ packages/suite/src/utils/suite/staking.ts
@@ -1,6 +1,9 @@
+import { Account, StakeType } from '@suite-common/wallet-types';
 import { BigNumber } from '@trezor/utils';
 
 import { TranslationFunction } from 'src/hooks/suite/useTranslation';
+
+import { getEthereumStakingAddressByType } from './ethereumStaking';
 
 interface ValidateMaxOptions {
     maxAmount: BigNumber;
@@ -21,4 +24,16 @@ export const calculateGains = (input: string, apy: number, divisor: number) => {
     const amount = new BigNumber(input).multipliedBy(apy).dividedBy(100).dividedBy(divisor);
 
     return amount.toFixed(5, 1);
+};
+
+export const getStakingContractAddress = (account: Account, stakeType: StakeType) => {
+    if (!account) return '';
+
+    switch (account.networkType) {
+        case 'ethereum':
+            return getEthereumStakingAddressByType(account.symbol, stakeType);
+        case 'solana':
+        default:
+            return account.descriptor;
+    }
 };
